move test_read_gbq_w_json* to test_session.py
